+++ tests/test_openapi_headers.py
@@ -1,3 +1,6 @@
+import openapi_spec_validator as osv
+import pytest
+
 from .schemas import Foo
 from .schemas import ResponseHeader
 from apiflask.fields import Boolean
@@ -25,36 +28,26 @@ def test_spec_with_dict_headers(app, client):
     rv = client.get('/openapi.json')
     assert rv.json['paths']['/foo']['get']['responses']['200']['headers'] == {
         'X-boolean': {
-            'name': 'X-boolean',
-            'in': 'header',
             'description': 'A boolean header',
             'required': False,
             'schema': {'type': 'boolean'},
         },
         'X-integer': {
-            'name': 'X-integer',
-            'in': 'header',
             'description': 'An integer header',
             'required': False,
             'schema': {'type': 'integer'},
         },
         'X-number': {
-            'name': 'X-number',
-            'in': 'header',
             'description': 'A number header',
             'required': False,
             'schema': {'type': 'number'},
         },
         'X-string': {
-            'name': 'X-string',
-            'in': 'header',
             'description': 'A string header',
             'required': False,
             'schema': {'type': 'string'},
         },
         'X-array': {
-            'name': 'X-array',
-            'in': 'header',
             'description': 'An array header',
             'required': False,
             'schema': {'items': {'type': 'string'}, 'type': 'array'},
@@ -83,10 +76,22 @@ def test_spec_with_schema_headers(app, client):
     rv = client.get('/openapi.json')
     assert rv.json['paths']['/foo']['get']['responses']['200']['headers'] == {
         'X-Token': {
-            'name': 'X-Token',
-            'in': 'header',
             'description': 'A custom token header',
             'required': True,
             'schema': {'type': 'string'},
         },
     }
+
+
+@pytest.mark.parametrize('openapi_version', ['3.0.0', '3.1.0'])
+def test_spec_validity_with_headers(app, client, openapi_version):
+    app.config['OPENAPI_VERSION'] = openapi_version
+
+    @app.route('/foo')
+    @app.output(Foo, headers=ResponseHeader)
+    def foo():
+        pass
+
+    rv = client.get('/openapi.json')
+    assert rv.status_code == 200
+    osv.validate(rv.json)