+++ CHANGES.md
@@ -1,3 +1,12 @@
+## Version 2.3.3
+
+Released: TODO
+
+- Convert the relative redirect path of OAuth2 to an absolute redirect URL. ([issue #602][issue_602]).
+
+[issue_602]: https://github.com/apiflask/apiflask/issues/602
+
+
 ## Version 2.3.2
 
 Released: 2024/12/15